[test] fix UI blocking issue
in UiTest::OnEnter, we setup an alarm for auto-leave a test item.

When the alarm fired, the UI thread walked to UiTest::OnLeave
and wait for the test thread done, so the input event cannot disptch to Ui thread.
but the test thread wait for the touch input event.

so, both of UI thread and test thread are blocking.

solution: before the alarm fire, we cancel it.